feat: version 3.0.0-beta, adapter rework, cleanup exports, reorg
