+++ src/interfaces/index.ts
@@ -1,0 +1,12 @@
+export * from './obj.before-fetch'
+export * from './obj.customer'
+export * from './obj.end-user'
+export * from './obj.file'
+export * from './obj.hooks'
+export * from './obj.interaction-event'
+export * from './obj.load-options'
+export * from './obj.meta'
+export * from './obj.record'
+export * from './obj.settings'
+export * from './obj.theme'
+export * from './obj.validation-response'
