+++ src/interfaces/obj.before-fetch.ts
