+++ src/interfaces/obj.hooks.ts
