+++ src/interfaces/obj.interaction-event.ts
